move onion creation to a separate library
