+++ onion/Cargo.toml
@@ -1,45 +1,29 @@
 [package]
-name = "mwixnet"
+name = "grin_onion"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[workspace]
-members = ["onion"]
 
 [dependencies]
 blake2 = { package = "blake2-rfc", version = "0.2"}
 byteorder = "1"
 bytes = "0.5.6"
 chacha20 = "0.8.1"
-clap = { version = "2.33", features = ["yaml"] }
 curve25519-dalek = "2.1"
-dirs = "2.0"
 ed25519-dalek = "1.0.1"
-function_name = "0.3.0"
-futures = "0.3"
 hmac = { version = "0.12.0", features = ["std"]}
-hyper = { version = "0.14", features = ["full"] }
-hyper-proxy = "0.9.1"
 itertools = { version = "0.10.3"}
-jsonrpc-core = "18.0"
-jsonrpc-derive = "18.0"
-jsonrpc-http-server = "18.0"
 lazy_static = "1"
-pbkdf2 = "0.8.0"
 rand = "0.7.3"
-ring = "0.16"
 rpassword = "4.0"
 serde = { version = "1", features= ["derive"]}
 serde_derive = "1"
 serde_json = "1"
 sha2 = "0.10.0"
 thiserror = "1.0.31"
-tokio = { version = "1", features = ["full"] }
 toml = "0.5"
 x25519-dalek = "0.6.0"
-grin_onion = { path = "./onion" }
 grin_secp256k1zkp = { version = "0.7.11", features = ["bullet-proof-sizing"]}
 grin_util = "5"
 grin_api = { git = "https://github.com/mimblewimble/grin", version = "5.2.0-alpha.1" }
