+++ onion/src/crypto/secp.rs
@@ -60,58 +60,3 @@ pub fn sign(sk: &SecretKey, msg: &Message) -> Result<Signature, secp256k1zkp::Er
 	let sig = aggsig::sign_single(&secp, &msg, &sk, None, None, None, Some(&pubkey), None)?;
 	Ok(sig)
 }
-
-#[cfg(test)]
-pub mod test_util {
-	use crate::crypto::secp::{self, Commitment, RangeProof, Secp256k1, SecretKey};
-	use grin_core::core::hash::Hash;
-	use grin_util::ToHex;
-	use rand::RngCore;
-
-	pub fn rand_commit() -> Commitment {
-		secp::commit(rand::thread_rng().next_u64(), &secp::random_secret()).unwrap()
-	}
-
-	pub fn rand_hash() -> Hash {
-		Hash::from_hex(secp::random_secret().to_hex().as_str()).unwrap()
-	}
-
-	pub fn rand_proof() -> RangeProof {
-		let secp = Secp256k1::new();
-		secp.bullet_proof(
-			rand::thread_rng().next_u64(),
-			secp::random_secret(),
-			secp::random_secret(),
-			secp::random_secret(),
-			None,
-			None,
-		)
-	}
-
-	pub fn proof(
-		value: u64,
-		fee: u32,
-		input_blind: &SecretKey,
-		hop_excesses: &Vec<&SecretKey>,
-	) -> (Commitment, RangeProof) {
-		let secp = Secp256k1::new();
-
-		let mut blind = input_blind.clone();
-		for hop_excess in hop_excesses {
-			blind.add_assign(&secp, &hop_excess).unwrap();
-		}
-
-		let out_value = value - (fee as u64);
-
-		let rp = secp.bullet_proof(
-			out_value,
-			blind.clone(),
-			secp::random_secret(),
-			secp::random_secret(),
-			None,
-			None,
-		);
-
-		(secp::commit(out_value, &blind).unwrap(), rp)
-	}
-}
