+++ onion/src/onion.rs
@@ -1,5 +1,4 @@
 use crate::crypto::secp::{self, Commitment, RangeProof, SecretKey};
-use crate::onion::OnionError::{InvalidKeyLength, SerializationError};
 use crate::util::{read_optional, vec_to_array, write_optional};
 
 use chacha20::cipher::{NewCipher, StreamCipher};
@@ -19,7 +18,7 @@ use thiserror::Error;
 use x25519_dalek::{PublicKey as xPublicKey, SharedSecret, StaticSecret};
 
 type HmacSha256 = Hmac<Sha256>;
-type RawBytes = Vec<u8>;
+pub type RawBytes = Vec<u8>;
 
 const CURRENT_ONION_VERSION: u8 = 0;
 
@@ -70,7 +69,6 @@ impl Payload {
 		Ok(payload)
 	}
 
-	#[cfg(test)]
 	pub fn serialize(&self) -> Result<Vec<u8>, ser::Error> {
 		let mut vec = vec![];
 		ser::serialize_default(&mut vec, &self)?;
@@ -166,7 +164,7 @@ impl Onion {
 	}
 }
 
-fn new_stream_cipher(shared_secret: &SharedSecret) -> Result<ChaCha20, OnionError> {
+pub fn new_stream_cipher(shared_secret: &SharedSecret) -> Result<ChaCha20, OnionError> {
 	let mut mu_hmac = HmacSha256::new_from_slice(b"MWIXNET")?;
 	mu_hmac.update(shared_secret.as_bytes());
 	let mukey = mu_hmac.finalize().into_bytes();
@@ -318,129 +316,21 @@ pub enum OnionError {
 
 impl From<InvalidLength> for OnionError {
 	fn from(_err: InvalidLength) -> OnionError {
-		InvalidKeyLength
+		OnionError::InvalidKeyLength
 	}
 }
 
 impl From<ser::Error> for OnionError {
 	fn from(err: ser::Error) -> OnionError {
-		SerializationError(err)
-	}
-}
-
-#[cfg(test)]
-pub mod test_util {
-	use super::{Onion, OnionError, Payload, RawBytes};
-	use crate::crypto::secp::test_util::{rand_commit, rand_proof};
-	use crate::crypto::secp::{random_secret, Commitment, SecretKey};
-
-	use chacha20::cipher::StreamCipher;
-	use grin_core::core::FeeFields;
-	use rand::{thread_rng, RngCore};
-	use secp256k1zkp::pedersen::RangeProof;
-	use x25519_dalek::PublicKey as xPublicKey;
-	use x25519_dalek::{SharedSecret, StaticSecret};
-
-	#[derive(Clone)]
-	pub struct Hop {
-		pub server_pubkey: xPublicKey,
-		pub excess: SecretKey,
-		pub fee: FeeFields,
-		pub rangeproof: Option<RangeProof>,
-	}
-
-	pub fn new_hop(
-		server_key: &SecretKey,
-		hop_excess: &SecretKey,
-		fee: u32,
-		proof: Option<RangeProof>,
-	) -> Hop {
-		Hop {
-			server_pubkey: xPublicKey::from(&StaticSecret::from(server_key.0.clone())),
-			excess: hop_excess.clone(),
-			fee: FeeFields::from(fee as u32),
-			rangeproof: proof,
-		}
-	}
-
-	/// Create an Onion for the Commitment, encrypting the payload for each hop
-	pub fn create_onion(commitment: &Commitment, hops: &Vec<Hop>) -> Result<Onion, OnionError> {
-		if hops.is_empty() {
-			return Ok(Onion {
-				ephemeral_pubkey: xPublicKey::from([0u8; 32]),
-				commit: commitment.clone(),
-				enc_payloads: vec![],
-			});
-		}
-
-		let mut shared_secrets: Vec<SharedSecret> = Vec::new();
-		let mut enc_payloads: Vec<RawBytes> = Vec::new();
-		let mut ephemeral_sk = StaticSecret::from(random_secret().0);
-		let onion_ephemeral_pk = xPublicKey::from(&ephemeral_sk);
-		for i in 0..hops.len() {
-			let hop = &hops[i];
-			let shared_secret = ephemeral_sk.diffie_hellman(&hop.server_pubkey);
-			shared_secrets.push(shared_secret);
-
-			ephemeral_sk = StaticSecret::from(random_secret().0);
-			let next_ephemeral_pk = if i < (hops.len() - 1) {
-				xPublicKey::from(&ephemeral_sk)
-			} else {
-				xPublicKey::from([0u8; 32])
-			};
-
-			let payload = Payload {
-				next_ephemeral_pk,
-				excess: hop.excess.clone(),
-				fee: hop.fee.clone(),
-				rangeproof: hop.rangeproof.clone(),
-			};
-			enc_payloads.push(payload.serialize()?);
-		}
-
-		for i in (0..shared_secrets.len()).rev() {
-			let mut cipher = super::new_stream_cipher(&shared_secrets[i])?;
-			for j in i..shared_secrets.len() {
-				cipher.apply_keystream(&mut enc_payloads[j]);
-			}
-		}
-
-		let onion = Onion {
-			ephemeral_pubkey: onion_ephemeral_pk,
-			commit: commitment.clone(),
-			enc_payloads,
-		};
-		Ok(onion)
-	}
-
-	pub fn rand_onion() -> Onion {
-		let commit = rand_commit();
-		let mut hops = Vec::new();
-		let k = (thread_rng().next_u64() % 5) + 1;
-		for i in 0..k {
-			let rangeproof = if i == (k - 1) {
-				Some(rand_proof())
-			} else {
-				None
-			};
-			let hop = new_hop(
-				&random_secret(),
-				&random_secret(),
-				thread_rng().next_u32(),
-				rangeproof,
-			);
-			hops.push(hop);
-		}
-
-		create_onion(&commit, &hops).unwrap()
+		OnionError::SerializationError(err)
 	}
 }
 
 #[cfg(test)]
 pub mod tests {
-	use super::test_util::{new_hop, Hop};
 	use super::*;
 	use crate::crypto::secp::random_secret;
+	use crate::{new_hop, Hop};
 
 	use grin_core::core::FeeFields;
