+++ onion/src/util.rs
